Update 4_Conditions_Conditional_Statements.ipynb
Noticed that when rendered on GitHub it didn't look the same as on my JupyterLab Notebook so I asked ChatGPT why and realized the html is too complicated so I went side by side and made the html less complex for the areas not rendering correctly. I have bold italics with *** on each side but GitHub doesn't render that so renders as *bold*, didn't go through all the code to fix that at this time.